configure pagination on mirage
