+++ src/pages/users/index.tsx
@@ -6,40 +6,12 @@ import { Header } from "../../components/Header";
 import { Pagination } from "../../components/Pagination";
 import { Sidebar } from "../../components/Sidebar";
 import Link from 'next/link'
-import { useEffect } from "react";
-import { useQuery } from "react-query"
-import { api } from "../../services/api";
+import { useUsers } from "../../services/hooks/useUsers";
 
-type UserProps = {
-
-    users: {
-        id: string;
-        name: string;
-        email: string;
-        createdAt: string;
-    }[]  
-}
 
 export default function UserList() {
 
-    const { data, isLoading, isFetching, error } = useQuery('users', async () => {
-        const { data } = await api.get<UserProps>("users")
-
-        const users = data.users.map(user => {
-            return {
-                ...user,
-                createdAt: new Date(user.createdAt).toLocaleDateString('pt-BR', {
-                    day: '2-digit',
-                    month: 'short',
-                    year: 'numeric'
-                })
-            }
-        })
-
-        return users
-    }, {
-        staleTime: 1000 * 5 // 5 seconds
-    })
+    const { data, isLoading, isFetching, error } =  useUsers()
 
     const isWideVersion = useBreakpointValue({
         base: false,
