+++ src/services/mirage/index.ts
@@ -1,4 +1,4 @@
-import { createServer, Factory, Model } from "miragejs"
+import { createServer, Factory, Model, Response } from "miragejs"
 
 import faker from "faker"
 
@@ -31,13 +31,33 @@ export function makeServer() {
         },
 
         seeds(server){
-            server.createList('user', 10)
+            server.createList('user', 200)
         },
 
         routes(){
             this.namespace = 'api'
             this.timing = 750; // ms
-            this.get('/users')
+            this.get('/users', function(schema, request) {
+                const { page = 1 , perPage = 10 } = request.queryParams;
+
+                const total = schema.all('user').length
+
+                const pageStart = (Number(page) - 1) * Number(perPage);
+                const pageEnd = pageStart + Number(perPage);
+
+                const users =  this.serialize(schema.all('user')).users.slice(pageStart, pageEnd)
+
+                return new Response(
+                    200,
+                    {
+                        'x-total-count': total.toString()
+                    }, 
+                    {
+                        users
+                    } 
+                )  
+                 
+            })
             this.post('/users')
             this.namespace = ''
             this.passthrough()
